@DorianZheng 2018-07-06
I have already run db_test and table_test for this PR, all tests passed.